Fix heap OOB read in WebGL texImage2D unpack-flip path
The UNPACK_FLIP_Y fast path allocated a tightly packed w*h*components
buffer via wgl_flip_rows, but glTexImage2D/glTexSubImage2D then read
wgl_transfer_bytes() worth of data, which honours UNPACK_ROW_LENGTH,
UNPACK_SKIP_PIXELS/ROWS and UNPACK_ALIGNMENT. With those pixel-store
params set, the GL transfer size exceeds the flipped allocation, so GL
reads past the end of the heap buffer — an attacker-controllable
out-of-bounds read whose contents leak back through readPixels.

Gate the in-tree flip on the transfer size equalling the tight size, so
the fast path is only taken when no pixel-store inflation applies; in all
other cases GL consumes the original, fully sized buffer.

Also harden two render-IPC size checks against signed overflow and free
a leaked EVP_PKEY on the WebCrypto JWK import error paths.

+++ src/rproc_http.c
@@ -525,7 +525,8 @@ ns_rproc_http_render(ns_rproc_http *r, int width, int height, int scroll_x,
         head.x_h < 1 || head.x_h > r->max_h ||
         head.x_stride < head.x_w * 4 || head.x_stride > (long)r->max_w * 4)
         return -1;
-    if (!r->shm && head.content_length < head.x_stride * head.x_h)
+    if (!r->shm && (uint64_t)head.content_length <
+            (uint64_t)head.x_stride * (uint64_t)head.x_h)
         return -1;
     out->ok = 1;
     out->width = (int)head.x_w;
@@ -993,7 +994,8 @@ ns_rproc_http_favicon(ns_rproc_http *r, int *out_w, int *out_h, int *out_stride)
     }
     if (head.x_w < 1 || head.x_w > 1024 || head.x_h < 1 || head.x_h > 1024 ||
         head.x_stride < head.x_w * 4 || head.x_stride > (long)1024 * 4 ||
-        head.x_stride * head.x_h > head.content_length) {
+        (uint64_t)head.x_stride * (uint64_t)head.x_h >
+            (uint64_t)head.content_length) {
         free(body);
         return NULL;
     }
